Worked on Product
Changed Product and Product List to make it easier for use to use - for
business case 3 and 5.  I was working on trying to get it working
through UI before stopping.  Need to know how user input will be handle
to complete UC3.

I will work on UC6 after short break and try to get it working.